backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
pytesseract==0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pydantic==2.5.0
PyPDF2==3.0.1
pdf2image==1.16.3
openpyxl>=3.1.0
python-dateutil>=2.8.0